+++ pgcrypto/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class PgcryptoConfig(AppConfig):
+    name = 'pgcrypto'
+
+    def ready(self):
+        """Ready for pgcrypto."""
+        pass