+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 2.3.0
+## 2.3.1
 
 * Added `__range` lookup for Date / DateTime fields (#59)
 * Remove compatibility for `Django 1.8, 1.9, and 1.10` (#62)
@@ -9,7 +9,10 @@
 * Added additional shields to `README`
 * Updated Travis config to include Python 3.5 and 3.6
 * Refactored lookups and mixins
-* Added support for `AppConfig` (#18)
+
+## 2.3.0
+
+* Invalid release, bump to 2.3.1
 
 ## 2.2.0